added seperate styles file for pomodoro timer
