+++ Aura_frontend/app/(tabs)/PomodoroScreen/PomodoroScreen.styles.ts
@@ -1,0 +1,140 @@
+import { StyleSheet } from "react-native";
+
+
+const styles = StyleSheet.create({
+  container: { flex: 1, backgroundColor: "#ffffff" },
+  scrollContent: { flexGrow: 1, paddingBottom: 150 },
+  taskTitle: { fontSize: 18, fontWeight: "bold", color: "#294C0D" },
+  modeSelector: {
+    flexDirection: "row",
+    justifyContent: "center",
+    marginVertical: 10,
+  },
+  modeButton: {
+    borderRadius: 10,
+    paddingVertical: 8,
+    paddingHorizontal: 16,
+    marginHorizontal: 5,
+    backgroundColor: "#eee",
+  },
+  activeMode: { backgroundColor: "#5FB21F" },
+  modeText: { color: "#294C0D", fontWeight: "bold" },
+  circleWrapper: {
+    alignSelf: "center",
+    marginTop: 20,
+    padding: 5,
+    borderRadius: 160,
+  },
+  circleInner: {
+    width: 200,
+    height: 200,
+    borderRadius: 100,
+    backgroundColor: "#fff",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  timerText: { fontSize: 36, fontWeight: "bold", color: "#294C0D" },
+  startButton: {
+    backgroundColor: "#5FB21F",
+    marginTop: 10,
+    paddingHorizontal: 20,
+    paddingVertical: 10,
+    borderRadius: 8,
+  },
+  startText: { color: "#fff", fontWeight: "bold", fontSize: 16 },
+  sectionTitle: {
+    fontSize: 18,
+    fontWeight: "bold",
+    marginBottom: 10,
+    paddingRight: 300,
+    color: "black",
+    borderBottomWidth: 3,
+    borderBottomColor: "black",
+    paddingBottom: 5,
+  },
+  taskItem: {
+    backgroundColor: "#fff",
+    padding: 12,
+    borderRadius: 10,
+    marginBottom: 10,
+    elevation: 2,
+    flexDirection: "column",
+  },
+  addTaskButton: {
+    borderStyle: "dashed",
+    borderWidth: 1,
+    borderColor: "#5FB21F",
+    padding: 10,
+    borderRadius: 8,
+    alignItems: "center",
+    marginTop: 10,
+  },
+  addTaskText: { color: "#5FB21F", fontWeight: "bold" },
+  modalOverlay: {
+    flex: 1,
+    backgroundColor: "rgba(0,0,0,0.5)",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  modalContent: {
+    width: "90%",
+    backgroundColor: "#fff",
+    borderRadius: 10,
+    padding: 20,
+  },
+  modalTitle: { fontSize: 18, fontWeight: "bold", marginBottom: 10 },
+  input: {
+    borderRadius: 8,
+    padding: 10,
+    marginBottom: 10,
+    fontStyle: "italic",
+  },
+  modalButton: {
+    flex: 1,
+    padding: 10,
+    marginHorizontal: 5,
+    borderRadius: 8,
+    alignItems: "center",
+  },
+  modalButtonText: { color: "#fff", fontWeight: "bold" },
+  currentTaskCard: {
+    backgroundColor: "#D1F2C8",
+    alignItems: "center",
+    marginHorizontal: 20,
+    marginTop: 20,
+    padding: 10,
+    borderRadius: 10,
+    elevation: 3,
+    shadowColor: "#000",
+    shadowOffset: { width: 0, height: 1 },
+    shadowOpacity: 0.2,
+    shadowRadius: 1.41,
+  },
+  currentTaskName: { fontSize: 16, fontWeight: "bold", color: "#294C0D" },
+  currentTaskMeta: { fontSize: 12, color: "#294C0D" },
+  menuContainer: {
+    backgroundColor: "#fff",
+    position: "absolute",
+    right: 20,
+    top: 30,
+    padding: 10,
+    borderRadius: 8,
+    elevation: 5,
+    zIndex: 10,
+  },
+  menuItem: { paddingVertical: 8, fontSize: 16, color: "#294C0D" },
+  taskItemContainer: { position: "relative", marginBottom: 10 },
+  taskMenu: {
+    position: "absolute",
+    top: 45,
+    right: 20,
+    backgroundColor: "#fff",
+    borderRadius: 8,
+    elevation: 5,
+    paddingLeft: 20,
+    paddingRight: 20,
+    zIndex: 10,
+  },
+});
+
+export default styles;
